[WIP] [Sema] Derive conformances to `KeyPathIterable` for classes.
The compiler now derives conformances `KeyPathIterable` for class types
based on the class's stored properties. Derivation of `allKeyPaths` works
just like it does for structs.

Conceptually, it makes sense for the implementation of `allKeyPaths` for subclasses
to prepend `super.allKeyPaths` if the superclass conforms to `KeyPathIterable`.

However, we found it impossible to express this in the type system because
`PartialKeyPath<BaseClass>` cannot be cast to `PartialKeyPath<SubClass>`.

Furthermore, subclasses inherit conformances from their superclass, so
`KeyPathIterable` synthesis will not be triggered for subclasses whose
superclass conforms to `KeyPathIterable`. This is a known phenomenon and
we don't attempt to derive conformance for such subclasses (`Codable`
synthesis has the same behavior).