[pmo] Use the linear lifetime checker to insert compensating destroys given only non-postdominating uses.
With ownership PMO needs to insert copies before the stores that provide an
available value. Usually, we are only using the available value along a single
path. This change ensures that if in that situation, we have "leaking paths"
besides that single path, PMO inserts compensating destroys.

NOTE: Without ownership enabled this is NFC.