Initial tracing support for functions (referred to as `tracee functions`) with the following signature: `(State, Data) -> (State, Result)`
This is designed to support the computation of a training step, where the
function takes and returns a Model (aka "state"), containing a collection of
trainable parameters. The function also takes as input a Data batch.

The additional output of `Result` can be used for the training step to produce
metrics (e.g. loss value).