AST: Don't use standard substitutions when mangling typealiases
Swift.UnicodeScalar is a typealias now, and for the TypeDecoder
to work correctly we need to mangle it as a typealias and not as
a struct.

However, we still want to demangle old names that used the
substitution, so don't remove it, instead just skip checking for
a substitution when mangling a typealias.

This does not change the ABI; typealiases are only mangled as
part of debug info.