Initial tracing support for functions (referred to as `tracee functions`)  (#22076)
* Initial tracing support for functions (referred to as `tracee functions`) with
the following signature: `(State, Data) -> (State, Result)`

This is designed to support the computation of a training step, where the
function takes and returns a Model (aka "state"), containing a collection of
trainable parameters. The function also takes as input a Data batch.

The additional output of `Result` can be used for the training step to produce
metrics (e.g. loss value).

* Addressed feedback.

* Addressed more feedback.

* Addressed more feedback and enhanced unit testing to run the created trace graph
function multiple times.

This requires that we finalize the trace graph function *before* running the
return closure of `trace()`, so that in that return closure, we are no longer in
tracing mode, and focus on executing the trace graph function.

* Addressed more feedback.

* Some more style tweaks.