Python 2/3 compat: update_checkout, build_swift, swift_build_support
Note: Unlike Python 2, in Python 3 `unittest discover` loads the module, forcing the manual inclusion of a known path to swift_build_support in update_checkout
Note: Renamed `ParserError` to `ArgParserError` to avoid confusion with `parser` module in build_swift
Note: The syntax for declaring a metaclass changed between Python 2 and 3 in a severely incompatible way; it is not possible to (sanely) use the Python 3 syntax with a Python 2 interpreter. Here we take inspiration from the six module's `with_metaclass()` function to directly specify the metaclass in a way that, while ugly, is fully cross-compatible.