Precheck SourceRange validity assertion
ASTVerifier forces various nodes to build SourceRanges out of their start and end SourceLocs; that can trip an assertion in SourceRange::SourceRange(), which of course doesn’t print any information about the SourceLocs in question. This change instead makes ASTVerifier try to check the precondition ahead of time so it can dump the AST node and SourceLocs if it’s violated. It won’t always work because some nodes implement getStartLoc()/getEndLoc() in terms of getSourceRange(), but when it does, it’ll give us more useful output.