[loadable-by-addr] Use undef rather than a copy_value to transfer an operand list from an original argument to a replaced argument.
Saves memory and eliminates a copy of a trivial value. I am tightening up the
invariant that copy_value, destroy_value can only get passed non-trivial values.

I don't think it is possible to write a test for this without my verification.
But I will be checking it in in a little bit.