[diagnose-unreachable] Emit a diagnostic if a Never field of a destructured struct is used.
Sometimes SILGen will emit a destructure_struct instead of struct_extract. So I
am broadening this check to handle destructure_struct to ensure that in those
cases, we still get an unreachable code diagnostic. The key thing here is that
we only emit the error if the Never field is actually used. This replicates the
behavior of the struct_extract since we can not destructure individual fields,
only the whole value.