Update WindowsBuild.md
Default to RelWithDebInfo builds on Windows because the build times with everything debug is ludicrous.  This gives a relatively useful build for development, even though the targets built with `clang-cl` have a slightly poorer debugging experience.  This reduces the build time by an order of magnitude though.  In theory, it is possible to build a second build of LLVM and Clang to get a debug build of the swift compiler building with a release build of clang.