Fix a race condition with the initialization of class metadata.
In the initial approach for resolving metadata dependency cycles,
superclass metadata was fetched (as non-transitively complete) by
the initialization function and passed to swift_initClassMetadata.
That was changed so that swift_initClassMetadata instead fetched
the superclass metadata via a demangling.  Unfortunately, it only
fetched *abstract* metadata, which is not actually safe to pass to
the ObjC runtime, which can manifest as a crash in either the runtime
or (potentially) downstream of it.  Worse, swift_initClassMetadata
has no way of reporting failure to its caller, and so we cannot
fix this problem without using a blocking fetch and subverting
the cyclic-metadata infrastructure.

The only way to solve this in the long term is to allow the runtime
to report failure via a metadata dependency from swift_initClassMetadata
(and swift_updateClassMetadata, which has a similar issue).
This patch is a naive version of that fix which changes the ABI of the
existing entrypoints to the ABI we need.  Because this is such a
late-breaking fix, this may no longer be possible, and we may need
to introduce new entrypoints with the right semantics.  If that is
necessary, I will restructure this patch.

Fixed rdar://47549859.